Elminate multiplications and divisions to do bit shifts
On EdgeTX 2.11, the floats shrunk from 64-bits to 32-bits.  A large 32-bit number cannot be stored in a 32-bit float, and data is lost,   giving bad results or in the case of the Activate, never get a complete activation code.

Additionally, cleanup some variables. Make them local to the function or to the script as needed.

All versions was changes to the same version but "a" as postfix to identify them